Expose thread_pool size as config option and decrease CRCON startup time (#187)
* Add configurable thread_pool size and skip connecting to the game server pre-emptively in rcon.cli

* Move types from rcon.extended_commands

* Move thread_pool stuff/team view down to RecordedRcon

* Update documentation for the thread pool size option

* Use RecordedRcon instead of Rcon due to thread pool changes

* Prevent warnings when no steam API key is set if the server is empty

* Fix player names not showing up in the audit message

* Move thread pool config to RecordedRcon

* Remove some debug stuff

* Add thread_pool_size constraint documentation

* Rename player_name parameter in do_message_player to match other commands
